Move model_config_loader to model_client
